renamed nagivator/comp to navigator
