+++ src/screens/Navigator/index.ts
@@ -1,9 +1,9 @@
-import { connect } from 'react-redux';
+import { connect } from "react-redux";
 
-import Navigator from './component';
-import { Item } from '../../types';
-import { setAmount } from '../../store/actions/amountLeft';
-import { setItems } from '../../store/actions/items';
+import Navigator from "./Navigator";
+import { Item } from "../../types";
+import { setAmount } from "../../store/actions/amountLeft";
+import { setItems } from "../../store/actions/items";
 
 const mapDispatchToProps = (dispatch: Function) => {
   return {
